emo-tracker/requirements.txt
beautifulsoup4==4.9.1
Brotli==1.0.7
bs4==0.0.1
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
dash==1.12.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
emoji==0.5.4
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
MarkupSafe==1.1.1
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
pandas==1.0.4
plotly==4.8.1
psycopg2-binary==2.8.5
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
threadpoolctl==2.1.0
tqdm==4.46.1
tweepy==3.8.0
urllib3==1.25.9
Werkzeug==1.0.1
